backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.7.0
cryptography>=41.0.0
aiohttp>=3.9.0
passlib==1.7.4
bcrypt==3.2.2

# Email (optional)
email-validator==2.1.0

# Redis (optional)
redis==5.0.1

# File handling
aiofiles==23.2.1
Pillow>=10.0.0
trimesh>=4.6.0
oss2>=2.18.0
alibabacloud_credentials>=0.3.6


# Utilities（pydantic 用 >=2.10 避免 Python 3.13 下需编译 pydantic-core）
pydantic>=2.10.0,<3
pydantic-settings>=2.1.0
python-dateutil==2.8.2

# SMS 短信（可选，用于手机验证码）
alibabacloud_dysmsapi20170525>=3.0.0
alibabacloud_tea_openapi>=0.3.0

# 本地加载 .env
python-dotenv>=1.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# API Key checking
colorama==0.4.6

# Payment
stripe==7.9.0
python-alipay-sdk>=3.0.0

# XSS 防护
bleach==6.1.0
